ready to merge

+++ lib/blocs/custom_dns_bloc/custom_dns_state.dart
@@ -6,6 +6,6 @@ class CustomDnsState with _$CustomDnsState {
   const factory CustomDnsState.loading() = _Loading;
   const factory CustomDnsState.loaded({
     required List<InternalCustomDNS> customDNSList,
-    required BigInt activatedDNSResolver,
+    required int activatedDNSResolver,
   }) = _Loaded;
 }
